requirements.txt
Flask==1.0.2
Flask-Session==0.3.1
Jinja2==2.10
mysql-connector-python-rf==2.2.2
request==1.0.2
requests==2.19.1

